Support exporting SwitchBot Hub2 metrics (temperature, humidity and light level).

+++ lib/src/main/kotlin/net/eplusx/mimosa/lib/switchbot/Hub2StatusResponse.kt
@@ -21,7 +21,7 @@ data class Hub2Status(
     val deviceType: String,
     val hubDeviceId: String,
     val version: String,
-    val temperature: Float,
+    val temperature: Double,
     val humidity: Int,
     val lightLevel: Int,
 )
